Fix : Entity ID int형에서 Long형으로 변경
JPA에서 ID 엔티티는 int 대신 Long형을 지원한다고하여 데이터베이스의 ID 엔티티 자료형을 Long으로 변경함.

+++ src/main/java/sch/cqre/api/model/entity/PostHashTagEntity.java
@@ -20,11 +20,11 @@ public class PostHashTagEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
 	@Column(name = "post_id")
-	private int postId;
+	private Long postId;
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
 	@Column(name = "hashtag_id")
-	private int hashtagId;
+	private Long hashtagId;
 
 	@Override
 	public boolean equals(Object o) {
